+++ tools/calc_xyz_transform.py
@@ -1,46 +1,59 @@
-"""Calculate XYZ conversion matrices."""
-import numpy as np
+"""
+Calculate XYZ conversion matrices.
 
-np.set_printoptions(precision=None, sign='-', floatmode='unique')
+Uses our own matrix tools for calculations, but can easily be replaced with `numpy`.
+"""
+import sys
+import os
+
+sys.path.insert(0, os.getcwd())
+
+from coloraide import util  # noqa: E402
+from coloraide import algebra as alg  # noqa: E402
+
+white_d65 = util.xy_to_xyz((0.31270, 0.32900))
+white_d50 = util.xy_to_xyz((0.34570, 0.35850))
 
 
-def xy_to_xyz(x, y):
-    """Convert `xyY` to `xyz`."""
-
-    return [x / y, 1, (1 - x - y) / y]
-
-
-white_d65 = np.asfarray(xy_to_xyz(0.31270, 0.32900))
-white_d50 = np.asfarray(xy_to_xyz(0.34570, 0.35850))
+def pprint(value):
+    """Print the matrix."""
+    print('[', end='')
+    first = True
+    for v in value:
+        if first:
+            first = False
+        else:
+            print(',\n ', end='')
+        print(v, end='')
+    print(']')
 
 
 def get_matrix(wp, space):
     """Get the matrices for the specified space."""
 
     if space == 'srgb':
-        x = np.asfarray([0.64, 0.30, 0.15])
-        y = np.asfarray([0.33, 0.60, 0.06])
+        x = [0.64, 0.30, 0.15]
+        y = [0.33, 0.60, 0.06]
     elif space == 'display-p3':
-        x = np.asfarray([0.68, 0.265, 0.150])
-        y = np.asfarray([0.32, 0.69, 0.060])
+        x = [0.68, 0.265, 0.150]
+        y = [0.32, 0.69, 0.060]
     elif space == 'rec2020':
-        x = np.asfarray([0.708, 0.17, 0.131])
-        y = np.asfarray([0.292, 0.797, 0.046])
+        x = [0.708, 0.17, 0.131]
+        y = [0.292, 0.797, 0.046]
     elif space == 'a98-rgb':
-        x = np.asfarray([0.64, 0.21, 0.15])
-        y = np.asfarray([0.33, 0.71, 0.06])
+        x = [0.64, 0.21, 0.15]
+        y = [0.33, 0.71, 0.06]
     elif space == 'prophoto-rgb':
-        x = np.asfarray([0.7347, 0.1596, 0.0366])
-        y = np.asfarray([0.2653, 0.8404, 0.0001])
+        x = [0.7347, 0.1596, 0.0366]
+        y = [0.2653, 0.8404, 0.0001]
     else:
         raise ValueError
 
-    one = np.float64(1.0)
-    m = np.transpose([xy_to_xyz(*xy) for xy in zip(x, y)])
-    mi = np.linalg.inv(m)
-    rgb = np.dot(mi, wp)
-    rgb2xyz = np.multiply(m, rgb)
-    xyz2rgb = np.linalg.inv(rgb2xyz)
+    m = alg.transpose([util.xy_to_xyz(xy) for xy in zip(x, y)])
+    mi = alg.inv(m)
+    rgb = alg.dot(mi, wp)
+    rgb2xyz = alg.multiply(m, rgb)
+    xyz2rgb = alg.inv(rgb2xyz)
 
     return rgb2xyz, xyz2rgb
 
@@ -49,34 +62,34 @@ if __name__ == "__main__":
     print('===== sRGB =====')
     to_xyz, from_xyz = get_matrix(white_d65, 'srgb')
     print('--- rgb -> xyz ---')
-    print(to_xyz)
+    pprint(to_xyz)
     print('--- xyz -> rgb ---')
-    print(from_xyz)
+    pprint(from_xyz)
 
     print('===== Display P3 =====')
     to_xyz, from_xyz = get_matrix(white_d65, 'display-p3')
     print('--- rgb -> xyz ---')
-    print(to_xyz)
+    pprint(to_xyz)
     print('--- xyz -> rgb ---')
-    print(from_xyz)
+    pprint(from_xyz)
 
     print('===== Adobe 98 =====')
     to_xyz, from_xyz = get_matrix(white_d65, 'a98-rgb')
     print('--- rgb -> xyz ---')
-    print(to_xyz)
+    pprint(to_xyz)
     print('--- xyz -> rgb ---')
-    print(from_xyz)
+    pprint(from_xyz)
 
     print('===== Rec.2020 =====')
     to_xyz, from_xyz = get_matrix(white_d65, 'rec2020')
     print('--- rgb -> xyz ---')
-    print(to_xyz)
+    pprint(to_xyz)
     print('--- xyz -> rgb ---')
-    print(from_xyz)
+    pprint(from_xyz)
 
     print('===== ProPhoto =====')
     to_xyz, from_xyz = get_matrix(white_d50, 'prophoto-rgb')
     print('--- rgb -> xyz ---')
-    print(to_xyz)
+    pprint(to_xyz)
     print('--- xyz -> rgb ---')
-    print(from_xyz)
+    pprint(from_xyz)